Refactor semian.c includes and types into header files
- Refactor semian includes into semian.h
- Refactor semian types into semian_types.h
